Refactored, added fabric8analytics.tests.usePersistantWorkspace property with profile
